Add EmailAddress class
