+++ packages/teams-js/src/public/meeting.ts
@@ -3,6 +3,7 @@ import { doesHandlerExist, registerHandler, removeHandler } from '../internal/ha
 import { ensureInitialized } from '../internal/internalAPIs';
 import { ApiName, ApiVersionNumber, getApiVersionTag } from '../internal/telemetry';
 import { FrameContexts } from './constants';
+import { EmailAddress } from './emailAddress';
 import { ErrorCode, SdkError } from './interfaces';
 import { runtime } from './runtime';
 
@@ -575,25 +576,6 @@ export namespace meeting {
      * the specified `contentUrl` passed to the {@link shareAppContentToStage} API in a new instance and screen share that instance.
      */
     ScreenShare = 'ScreenShare',
-  }
-
-  /**
-   * Represents a validated email.
-   *
-   * @hidden
-   * Hide from docs.
-   */
-  export class EmailAddress {
-    public constructor(private readonly val: string) {
-      const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-      if (!emailRegex.test(this.val)) {
-        throw new Error('Input email address does not have the correct format.');
-      }
-    }
-
-    public toString(): string {
-      return this.val;
-    }
   }
 
   /**
